Update component settings bar to respect dependsOn metadata
